# docs/openapi.yaml
openapi: 3.1.0
info:
  title: CaReMind API
  version: 1.0.0
  description: REST API for private vehicle, maintenance and operating-cost management.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: https://api.car-remind.gr/api
    description: Production
...
paths:
...
  /account/me:
    get:
      operationId: getCurrentAccount
      tags: [Account]
      summary: Read the current profile
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Profile }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /account/send-code:
    post:
      operationId: sendAccountCode
      tags: [Account]
      summary: Send an account-change verification code
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Code sent }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /account/verify-code:
    post:
      operationId: verifyAccountCode
      tags: [Account]
      summary: Exchange an account-change code for a short-lived token
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Account token created }
        '400': { $ref: '#/components/responses/BadRequest' }
  /account/update:
    post:
      operationId: updateAccount
      tags: [Account]
      summary: Update email, username or password with an account token
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Account updated }
        '400': { $ref: '#/components/responses/BadRequest' }
  /account/recipients:
    get:
      operationId: listNotificationRecipients
      tags: [Account]
      summary: List reminder recipients
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Recipient list }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createNotificationRecipient
      tags: [Account]
      summary: Add a reminder recipient
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Recipient added }
        '400': { $ref: '#/components/responses/BadRequest' }
  /account/recipients/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    delete:
      operationId: deleteNotificationRecipient
      tags: [Account]
      summary: Remove an owned reminder recipient
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Recipient removed }
        '404': { $ref: '#/components/responses/NotFound' }
  /users:
    get:
      operationId: listUsers
      tags: [Administration]
      summary: List users
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: User list }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      operationId: createUser
      tags: [Administration]
      summary: Create a user and company transactionally
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: User created }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    put:
      operationId: updateUser
      tags: [Administration]
      summary: Update a user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: User updated }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteUser
      tags: [Administration]
      summary: Delete a non-admin user
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: User deleted }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/{id}/toggle-active:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch:
      operationId: toggleUserStatus
      tags: [Administration]
      summary: Toggle a non-admin account status
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Status changed }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: integer, minimum: 1 }
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired authentication
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found or not owned by this user
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties: { error: { type: string }, code: { type: string } }
...